JL8-24 refactor: 프로필에딧 체인지폼 분리
